refactor modes into separate folder
