+++ examples/worf-warden/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2024"
 [dependencies]
 worf = {path = "../../worf"}
 env_logger = "0.11.8"
+log = "0.4.27"
 
 # todo re-add this
 #[features]
